Consolidate nxutil into dace.sdfg.utils
