+++ dace/frontend/tensorflow/transformations/redundant_array.py
@@ -1,7 +1,8 @@
 """ Contains classes that implement a redundant array removal transformation.
 """
 
-from dace.graph import nodes, nxutil
+from dace.graph import nodes
+from dace.sdfg import utils as sdutil
 from dace.transformation import pattern_matching as pm
 from dace.config import Config
 
@@ -17,7 +18,7 @@ class TensorflowRedundantArray(pm.Transformation):
     @staticmethod
     def expressions():
         return [
-            nxutil.node_path_graph(TensorflowRedundantArray._in_array,
+            sdutil.node_path_graph(TensorflowRedundantArray._in_array,
                                    TensorflowRedundantArray._out_array)
         ]
