+++ dace/transformation/dataflow/map_fission.py
@@ -3,7 +3,8 @@
 from copy import deepcopy as dcpy
 from collections import defaultdict
 from dace import registry, sdfg as sd, memlet as mm, subsets, data as dt
-from dace.graph import nodes, nxutil
+from dace.graph import nodes
+from dace.sdfg import utils as sdutil
 from dace.graph.graph import OrderedDiGraph
 from dace.symbolic import pystr_to_symbolic
 from dace.transformation import pattern_matching, helpers
@@ -41,8 +42,8 @@ class MapFission(pattern_matching.Transformation):
     @staticmethod
     def expressions():
         return [
-            nxutil.node_path_graph(MapFission._map_entry, ),
-            nxutil.node_path_graph(
+            sdutil.node_path_graph(MapFission._map_entry, ),
+            sdutil.node_path_graph(
                 MapFission._map_entry,
                 MapFission._nested_sdfg,
             )
