+++ dace/transformation/dataflow/mpi.py
@@ -2,7 +2,8 @@
 
 from dace import dtypes, registry
 from dace.sdfg import has_dynamic_map_inputs
-from dace.graph import nodes, nxutil
+from dace.sdfg import utils as sdutil
+from dace.graph import nodes
 from dace.transformation import pattern_matching
 from dace.properties import make_properties
 
@@ -57,7 +58,7 @@ class MPITransformMap(pattern_matching.Transformation):
 
     @staticmethod
     def expressions():
-        return [nxutil.node_path_graph(MPITransformMap._map_entry)]
+        return [sdutil.node_path_graph(MPITransformMap._map_entry)]
 
     @staticmethod
     def can_be_applied(graph, candidate, expr_index, sdfg, strict=False):
