+++ dace/transformation/interstate/fpga_transform_state.py
@@ -2,7 +2,8 @@
 
 import dace
 from dace import data, memlet, dtypes, registry, sdfg as sd, subsets
-from dace.graph import nodes, nxutil
+from dace.graph import nodes
+from dace.sdfg import utils as sdutil
 from dace.transformation import pattern_matching
 
 
@@ -35,7 +36,7 @@ class FPGATransformState(pattern_matching.Transformation):
 
     @staticmethod
     def expressions():
-        return [nxutil.node_path_graph(FPGATransformState._state)]
+        return [sdutil.node_path_graph(FPGATransformState._state)]
 
     @staticmethod
     def can_be_applied(graph, candidate, expr_index, sdfg, strict=False):
@@ -126,8 +127,8 @@ class FPGATransformState(pattern_matching.Transformation):
         state = sdfg.nodes()[self.subgraph[FPGATransformState._state]]
 
         # Find source/sink (data) nodes
-        input_nodes = nxutil.find_source_nodes(state)
-        output_nodes = nxutil.find_sink_nodes(state)
+        input_nodes = sdutil.find_source_nodes(state)
+        output_nodes = sdutil.find_sink_nodes(state)
 
         fpga_data = {}
 
@@ -194,11 +195,11 @@ class FPGATransformState(pattern_matching.Transformation):
 
                 if node not in wcr_input_nodes:
                     fpga_node = state.add_read('fpga_' + node.data)
-                    nxutil.change_edge_src(state, node, fpga_node)
+                    sdutil.change_edge_src(state, node, fpga_node)
                     state.remove_node(node)
 
             sdfg.add_node(pre_state)
-            nxutil.change_edge_dest(sdfg, state, pre_state)
+            sdutil.change_edge_dest(sdfg, state, pre_state)
             sdfg.add_edge(pre_state, state, sd.InterstateEdge())
 
         if output_nodes:
@@ -238,11 +239,11 @@ class FPGATransformState(pattern_matching.Transformation):
                 post_state.add_edge(post_fpga_node, None, post_node, None, mem)
 
                 fpga_node = state.add_write('fpga_' + node.data)
-                nxutil.change_edge_dest(state, node, fpga_node)
+                sdutil.change_edge_dest(state, node, fpga_node)
                 state.remove_node(node)
 
             sdfg.add_node(post_state)
-            nxutil.change_edge_src(sdfg, state, post_state)
+            sdutil.change_edge_src(sdfg, state, post_state)
             sdfg.add_edge(state, post_state, sd.InterstateEdge())
 
         veclen_ = 1
